Fix coverity issues 65866 and 65865 (#361)
65865: cast max_metrics_sample to 64bit to force operation to be evaluated using 64-bit arithmetic
65866: initialize both is_gpu and is_cpu to false so that there is a path the code that handles the case when neither are set on the command line

+++ sdk/src/metrics_handler.h
@@ -142,7 +142,7 @@ class PtiMetricsProfiler {
     // 2 is for systems with 2 tiles
     // 512 is for extra padding
     // TODO: may need to be adjusted per specific target
-    return max_metric_samples_ * PTI_METRIC_COUNT_MAX * 2 + 512;
+    return static_cast<uint64_t>(max_metric_samples_) * PTI_METRIC_COUNT_MAX * 2 + 512;
   }
 
   virtual ~PtiMetricsProfiler() {
